Refine OpenCL target resource lowering
