+++ tests/test_backend/test_OpenCL/test_external_fixtures.py
@@ -682,7 +682,7 @@ def test_external_opencv_opencl_const_parameter_codegen_reparse():
         "type": "__global__ const float *",
         "name": "src",
     }
-    assert "var<storage, read_write> src: array<f32>" in crossgl
+    assert "var<storage, read> src: array<f32>" in crossgl
     assert "var<workgroup> smem: array<f32>" in crossgl
     assert "var<storage, read_write> smem: array<f32>" not in crossgl
     assert "__const" not in crossgl
